Define types for "apigility", "expressive", and "zf" applications (#865)
Adds the script `serve-zf.sh`, which sets up an nginx host for use with
any of Apigility, Expressive, or zend-mvc applications, as they all use
the same setup.

In each case, developers should do the following in `Homestead.yaml`:

- create a `folders` mapping that maps the application root directory to
  a directory in the vagrant box.
- add a site that
  - defines a "to" setting that appends "/public" to the "to" mapping from the above step.
  - uses a "type" of either:
    - "apigility"
    - "expressive"
    - "zf"

The patch updates `scripts/homestead.rb` to add a `case` statement that
redefines the `type` if either `apigility`, `expressive`, or `symfony`
are detected. In the first two instances, they map to `zf`; the last
maps to `symfony2` (which it did previously via a conditional).