set "mysql" as an enabled database (#1925)
MySQL is kind of a weird case, because it's our default database, and it cannot (shouldn't?) run side by side with MariaDB.

We kind of cheated this previously by having "mysql" as a key in the "features" configuration, but it's technically not a "feature" as it doesn't have a file in `/scripts/features/`.  this resulted in an "Invalid feature: mysql" warning in the provision script.

This change will allow us to remove `mysql: true` from our docs and example files, while still having MySQL work correctly.